refactor: remove tree and move function to each component
