+++ src/components/SideMenu.js
@@ -1,5 +1,3 @@
-import { buildTree } from "../utils/tree.js";
-
 export default function SideMenu({
   $target,
   initialState,
@@ -72,3 +70,18 @@ export default function SideMenu({
     }
   });
 }
+
+const buildTree = (arr, targetId, depth) =>
+  `<ul data-depth=${depth}>${arr
+    .map(
+      ({ id, title, documents }) =>
+        `<li>${
+          documents.length
+            ? `<label><ion-icon name="chevron-down-outline" class="arrow"/></label>`
+            : "- "
+        }<span class="documentTitle ${
+          Number(targetId) === id ? "selected" : ""
+        }" data-id="${id}">${title}</span><button data-add="${id}"> + </button><button data-remove="${id}"> - </button>
+          </li>${buildTree(documents, targetId, depth + 1)}`
+    )
+    .join("")}</ul>`;
